Goddammit blackfriday
These guys have been sitting with a breaking unreleased API on `master`
for months so if you're referencing system packages you get an interface
that only supports `MarkdownCommon`, but if you use dep and it fetches a
released version, only `Run` is supported.

Now that we've vendored in dependencies, switch back to `Run`.